rename data/queries.js to client.js
